Support special characters in [JsonPropertyName]
