+++ src/libraries/System.Text.Json/gen/ContextGenerationSpec.cs
@@ -32,9 +32,11 @@ namespace System.Text.Json.SourceGeneration
         public HashSet<TypeGenerationSpec> TypesWithMetadataGenerated { get; } = new();
 
         /// <summary>
-        /// Cache of runtime property names (statically determined) found accross the object graph of the JsonSerializerContext.
+        /// Cache of runtime property names (statically determined) found across the object graph of the JsonSerializerContext.
+        /// The dictionary Key is the JSON property name, and the Value is the variable name which is the same as the property
+        /// name except for cases where special characters are used with [JsonPropertyName].
         /// </summary>
-        public HashSet<string> RuntimePropertyNames { get; } = new();
+        public Dictionary<string, string> RuntimePropertyNames { get; } = new();
 
         public string ContextTypeRef => ContextType.GetCompilableName();
     }
