+++ src/libraries/System.Text.Json/gen/PropertyGenerationSpec.cs
@@ -32,6 +32,8 @@ namespace System.Text.Json.SourceGeneration
         /// </summary>
         public string RuntimePropertyName { get; init; }
 
+        public string PropertyNameVarName { get; init; }
+
         /// <summary>
         /// Whether the property has a set method.
         /// </summary>
